feat/Add amplitude
