+++ package.json
@@ -11,6 +11,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@amplitude/analytics-react-native": "^1.4.11",
     "@notifee/react-native": "^9.1.3",
     "@react-native-community/geolocation": "^3.4.0",
     "@react-native-community/push-notification-ios": "^1.11.0",
